cloud-function/requirements.txt
google-cloud-storage==2.8.0
google-api-python-client==2.70.0
google-auth==2.15.0
pytz==2023.3